requirements.txt
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
speedtest-cli==2.1.3
tzdata==2024.2
